# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Linkedin-Hyper API
  version: 1.0.0
  description: >-
    Browser-facing Next.js BFF API for Linkedin-Hyper. This spec documents the
    operator dashboard routes, account/session flows, inbox APIs, send/sync APIs,
    health endpoints, export endpoints, and supporting account/stats endpoints.
servers:
  - url: /
    description: Same-origin frontend deployment
  - url: http://127.0.0.1:3002
    description: Local frontend container or dev server
  - url: http://localhost:3000
    description: Local Next.js dev server
...
paths:
...
  /api/accounts:
    get:
      tags: [Accounts]
      summary: List configured and known accounts
      security:
        - cookieAuth: []
        - serviceToken: []
      responses:
        '200':
          description: Accounts returned
          content:
            application/json:
              schema:
                type: object
                required: [accounts]
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/accounts/{id}/session:
    post:
      tags: [Sessions/Cookies]
      summary: Import LinkedIn session cookies for an account
      description: >-
        Accepts a non-empty cookie array, validates required LinkedIn auth
        cookies, and forwards the import to the worker.
      security:
        - cookieAuth: []
        - apiKeyAuth: []
        - serviceToken: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/CookieObject'
      responses:
        '200':
          description: Cookies imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookieImportResponse'
        '400':
          description: Invalid cookie payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Sessions/Cookies]
      summary: Delete imported session cookies for an account
      security:
        - cookieAuth: []
        - apiKeyAuth: []
        - serviceToken: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '204':
          description: Session deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/accounts/{id}/session/status:
    get:
      tags: [Sessions/Cookies]
      summary: Get stored session status for an account
      security:
        - cookieAuth: []
        - apiKeyAuth: []
        - serviceToken: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: Session exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatusResponse'
        '404':
          description: No stored session
          content:
            application/json:
              schema:
                type: object
                required: [exists]
                properties:
                  exists:
                    type: boolean
                    enum: [false]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/accounts/{id}/limits:
    get:
      tags: [Accounts]
      summary: Get rate limit state for an account
      security:
        - cookieAuth: []
        - serviceToken: []
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: Limit counters returned
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/connections/unified:
    get:
      tags: [Connections]
      summary: Get unified connections data
      security:
        - cookieAuth: []
        - serviceToken: []
      responses:
        '200':
          description: Connections payload returned
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/connections/send:
    post:
      tags: [Connections]
      summary: Send a connection request
      security:
        - cookieAuth: []
        - serviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountId, profileUrl]
              properties:
                accountId:
                  type: string
                profileUrl:
                  type: string
                note:
                  type: string
                  maxLength: 300
      responses:
        '200':
          description: Connection request forwarded to worker
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: app_session
      description: Dashboard session cookie.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Controlled operator path for specific worker/admin recovery routes only.
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Compatibility-only service bearer token path for operator automation.
  parameters:
    AccountIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Managed LinkedIn account identifier.
...
  responses:
    Unauthorized:
      description: Request is not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Request is authenticated but not authorized or same-origin checks failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          nullable: true
        retryAfterSec:
          type: integer
          nullable: true
...
    Account:
      type: object
      required: [id, displayName, isActive]
      properties:
        id:
          type: string
        displayName:
          type: string
        isActive:
          type: boolean
        lastSeen:
          type: string
          nullable: true
    CookieObject:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
        domain:
          type: string
        path:
          type: string
        expires:
          type: number
        secure:
          type: boolean
        httpOnly:
          type: boolean
        sameSite:
          type: string
      additionalProperties: true
    CookieImportResponse:
      type: object
      required: [success, accountId, cookieCount, message]
      properties:
        success:
          type: boolean
        accountId:
          type: string
        cookieCount:
          type: integer
        message:
          type: string
    SessionStatusResponse:
      type: object
      required: [exists]
      properties:
        exists:
          type: boolean
        savedAt:
          type: string
          nullable: true
        cookieCount:
          type: integer
          nullable: true
        requiredCookiesPresent:
          type: boolean
          nullable: true
      additionalProperties: true
...